move create field to RPF
